// src/app/p-app/p-purchase/shared/dto/DTOPartnerContact.dto.ts
export class DTOPartnerContact {
    Code: number = 0
    Company: number = 1
    COPartner: number = 0
    IsOversea: boolean = false
    ContactName: string = ""
    PositionName: string = ""
    Department: string = ""
    Tel1: string = ""
    Tel2: string = ""
    Ext1: string = ""
    Ext2: string = ""
    Cellphone1: string = ""
    Cellphone2: string = ""
    Email1: string = ""
    Email2: string = ""
    StatusName: string = ""
}